+++ examples/react-native/app.test.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import App from './app';
-
-import renderer from 'react-test-renderer';
-
-it('renders without crashing', () => {
-  const rendered = renderer.create(<App />).toJSON();
-  expect(rendered).toBeTruthy();
-});